user-service/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
ecdsa==0.19.1
fastapi==0.115.12
google-auth==2.38.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.4.8
pyasn1_modules==0.4.1
pydantic==2.11.3
pydantic_core==2.33.1
python-jose==3.4.0
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.0
